Split up the non-matrix f32 arithmetic tests
This reduces the size of the cache JSON that is needed to be loaded.

Issue #2604
